+++ crates/evm/src/block/state_changes.rs
@@ -26,14 +26,22 @@ pub fn post_block_balance_increments<H>(
 where
     H: BlockHeader,
 {
-    let mut balance_increments = HashMap::default();
+    let mut balance_increments = HashMap::with_capacity_and_hasher(
+        withdrawals.map_or(ommers.len(), |w| w.len()),
+        Default::default(),
+    );
 
     // Add block rewards if they are enabled.
-    if let Some(base_block_reward) = calc::base_block_reward(&spec, block_env.number) {
+    if let Some(base_block_reward) =
+        calc::base_block_reward(&spec, block_env.number.saturating_to())
+    {
         // Ommer rewards
         for ommer in ommers {
-            *balance_increments.entry(ommer.beneficiary()).or_default() +=
-                calc::ommer_reward(base_block_reward, block_env.number, ommer.number());
+            *balance_increments.entry(ommer.beneficiary()).or_default() += calc::ommer_reward(
+                base_block_reward,
+                block_env.number.saturating_to(),
+                ommer.number(),
+            );
         }
 
         // Full block reward
@@ -44,7 +52,7 @@ where
     // process withdrawals
     insert_post_block_withdrawals_balance_increments(
         spec,
-        block_env.timestamp,
+        block_env.timestamp.saturating_to(),
         withdrawals.map(|w| w.as_slice()),
         &mut balance_increments,
     );
@@ -122,6 +130,7 @@ where
                 info: account.info.clone(),
                 storage: Default::default(),
                 status: AccountStatus::Touched,
+                transaction_id: 0,
             },
         ))
     };